Add CSS file back in
Turns out that the website was just taking an unusually long time to update for some reason. I removed the CSS file entirely and my website still looks like it has perfect CSS, so that isn't updating either. I'm just going to add in the CSS file and wait a little while to see if it sticks.